Splits the Evaluation class
There is an assumption in the original class that the number of
cases must be the same for both the prediction and ground truth
sets. This is not the case for Detection evaluations where
we will probably have many more predictions than ground truth
cases. So, create a class for each.

TODO
* handle the _cases creation